API call 수정

+++ src/main/java/com/carnival/matchcube/dto/AreaDTO.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AreaDTO {
-    public int code;
+    public Long code;
 }
